app/requirements.txt

chromadb==0.5.18

langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-groq==0.1.10
langchain-text-splitters==0.2.4


pandas==2.0.3

python-dotenv==1.0.1

selenium==4.26.1

streamlit==1.40.1

uuid==1.30
